dicing up the cli
