+++ parser-typechecker/src/Unison/Codebase/CommandLine.hs
@@ -6,6 +6,8 @@
 {-# LANGUAGE TupleSections       #-}
 {-# LANGUAGE TypeApplications    #-}
 {-# LANGUAGE ViewPatterns        #-}
+{-# LANGUAGE DoAndIfThenElse     #-}
+
 
 module Unison.Codebase.CommandLine where
 
@@ -15,15 +17,10 @@ import           Prelude                 hiding ( readFile
                                                 )
 import           Control.Applicative            ((<|>))
 import           Control.Concurrent             (forkIO, killThread)
-import qualified Control.Concurrent.Async       as Async
 import           Control.Concurrent.STM         (atomically)
-import           Control.Exception              (finally)
 import           Control.Monad                  (forever, join, when)
-import           Control.Monad.IO.Class         (MonadIO, liftIO)
-import           Control.Monad.Trans            (lift)
 import           Data.Foldable                  (toList, traverse_)
-import           Data.IORef
-import           Data.List                      (intercalate, isSuffixOf, sort)
+import           Data.List                      (isSuffixOf, sort)
 import           Data.ListLike                  (ListLike)
 import           Data.List.Extra                (nubOrdOn)
 import           Data.Map                       (Map)
@@ -44,12 +41,11 @@ import           Unison.Codebase                (Codebase)
 import qualified Unison.Codebase                as Codebase
 import           Unison.Codebase.Branch         (Branch)
 import qualified Unison.Codebase.Branch         as Branch
+import           Unison.Codebase.CommandLine.InputPattern (InputPattern(parse))
 import           Unison.Codebase.Editor         (BranchName, DisplayThing (..),
                                                  Event (..), Input (..),
                                                  Output (..))
 import qualified Unison.Codebase.Editor         as E
-import qualified Unison.Codebase.Editor.Actions as Actions
-import           Unison.Codebase.Runtime        (Runtime)
 import qualified Unison.Codebase.Runtime        as Runtime
 import qualified Unison.Codebase.Watch          as Watch
 import qualified Unison.HashQualified           as HQ
@@ -88,8 +84,37 @@ notifyUser :: forall v . Var v => FilePath -> Output v -> IO ()
 notifyUser dir o = case o of
   Success (MergeBranchI _) -> putPrettyLn $ P.bold "Merged. " <> "Here's what's `todo` after the merge:"
   Success _    -> putPrettyLn $ P.bold "Done."
-  DisplayDefinitions outputLoc ppe terms types -> let
+  DisplayDefinitions outputLoc ppe terms types -> case outputLoc of
+    -- stdout
+    Nothing -> putPrettyLn out
+    Just path -> do
+      path' <- canonicalizePath path
+      prependToFile out path'
+      printMessage out path'
+      where
+      prependToFile out path = do
+        existingContents <- do
+          exists <- doesFileExist path
+          if exists then readFile path
+          else pure ""
+        writeFile path . Text.pack . P.toPlain 80 $
+          P.lines [ out, ""
+                  , "---- " <> "Anything below this line is ignored by Unison."
+                  , "", P.text existingContents ]
+      printMessage out path =
+        putPrettyLn . P.callout "☝️" $ P.lines [
+          P.wrap $ "I added these definitions to the top of " <> fromString path,
+          "",
+          P.indentN 2 out,
+          "",
+          P.wrap $
+            "You can edit them there, then do `update` to replace the" <>
+            "definitions currently in this branch."
+         ]
+    where
+    out = P.sep "\n\n" (prettyTypes <> prettyTerms)
     prettyTerms = map go terms
+    prettyTypes = map go2 types
     go (r, dt) =
       let n = PPE.termName ppe (Referent.Ref r) in
       case dt of
@@ -97,14 +122,6 @@ notifyUser dir o = case o of
         BuiltinThing -> builtin n
         RegularThing tm -> P.map fromString $
           TermPrinter.prettyBinding ppe n tm
-    prettyTypes = map go2 types
-    builtin n = P.wrap $ "--" <> prettyHashQualified n <> " is built-in."
-    missing n r = P.wrap (
-      "-- The name " <> prettyHashQualified n <> " is assigned to the "
-      <> "reference " <> fromString (show r ++ ",")
-      <> "which is missing from the codebase.")
-      <> P.newline
-      <> tip "You might need to repair the codebase manually."
     go2 (r, dt) =
       let n = PPE.typeName ppe r in
       case dt of
@@ -113,30 +130,13 @@ notifyUser dir o = case o of
         RegularThing decl -> case decl of
           Left _ability -> TypePrinter.prettyEffectHeader n <> " -- todo"
           Right _d -> TypePrinter.prettyDataHeader n <> " -- todo"
-    out = P.sep "\n\n" (prettyTypes <> prettyTerms)
-    in
-      case outputLoc of
-         Nothing -> putPrettyLn out
-         Just path -> do
-           path' <- canonicalizePath path
-           exists <- doesFileExist path'
-           existingContents <-
-             if exists then readFile path'
-             else pure ""
-           writeFile path' . Text.pack . P.toPlain 80 $
-             P.lines [ out, ""
-                     , "---- " <> "Anything below this line is ignored by Unison."
-                     , "", P.text existingContents ]
-           putPrettyLn . P.callout "☝️" $
-             P.lines [
-               P.wrap $ "I added these definitions to the top of " <> fromString path',
-               "",
-               P.indentN 2 out,
-               "",
-               P.wrap $
-                 "You can edit them there, then do `update` to replace the" <>
-                 "definitions currently in this branch."
-             ]
+    builtin n = P.wrap $ "--" <> prettyHashQualified n <> " is built-in."
+    missing n r = P.wrap (
+      "-- The name " <> prettyHashQualified n <> " is assigned to the "
+      <> "reference " <> fromString (show r ++ ",")
+      <> "which is missing from the codebase.")
+      <> P.newline
+      <> tip "You might need to repair the codebase manually."
   NoUnisonFile -> do
     dir' <- canonicalizePath dir
     putPrettyLn . P.callout "😶" $ P.lines
@@ -144,13 +144,12 @@ notifyUser dir o = case o of
       , ""
       , P.column2 [(P.bold "Hint:", msg dir')] ]
    where
-    msg dir =
-      P.wrap
-        $  "I'm currently watching for definitions in .u files under the"
-        <> renderFileName dir
-        <> "directory. Make sure you've updated something there before using the"
-        <> P.bold "`add`" <> "or" <> P.bold "`update`"
-        <> "commands."
+    msg dir = P.wrap
+      $  "I'm currently watching for definitions in .u files under the"
+      <> renderFileName dir
+      <> "directory. Make sure you've updated something there before using the"
+      <> P.bold "`add`" <> "or" <> P.bold "`update`"
+      <> "commands."
   UnknownBranch branchName ->
     putPrettyLn
       .  warn
@@ -233,7 +232,12 @@ notifyUser dir o = case o of
         formatMissingStuff termsWithMissingTypes missingTypes
       when (not $ null impossible) . error $ "Compiler bug, these referents are missing types: " <> show impossible
 
-  SlurpOutput s -> let
+  SlurpOutput s ->
+    putPrettyLn . P.sep "\n" . catMaybes $ [
+        addMsg, updateMsg, dupeMsg, collMsg,
+        conflictMsg, aliasingMsg, termExistingCtorMsg,
+        ctorExistingTermMsg, blockedDependenciesMsg ]
+    where
     -- todo: move this to a separate function
     branch = E.updatedBranch s
     file = E.originalFile s
@@ -400,10 +404,6 @@ notifyUser dir o = case o of
               TypePrinter.prettySignatures' ppe (filterTermTypes blockedTerms)
             )
            )
-    in putPrettyLn . P.sep "\n" . catMaybes $ [
-        addMsg, updateMsg, dupeMsg, collMsg,
-        conflictMsg, aliasingMsg, termExistingCtorMsg,
-        ctorExistingTermMsg, blockedDependenciesMsg ]
   ParseErrors src es -> do
     Console.setTitle "Unison ☹︎"
     traverse_ (putStrLn . CT.toANSI . prettyParseError (Text.unpack src)) es
@@ -420,14 +420,15 @@ notifyUser dir o = case o of
       watchPrinter fileContents ppe ann evald isCacheHit |
       (_v, (ann,evald,isCacheHit)) <- Map.toList watches ]
   DisplayConflicts branch -> do
-    let terms    = R.dom $ Branch.termNamespace branch
-        types    = R.dom $ Branch.typeNamespace branch
-    when (not $ null terms) $ do
-      putStrLn "🙅 These terms have conflicts: "
-      traverse_ (\x -> putStrLn ("  " ++ Name.toString x)) terms
-    when (not $ null types) $ do
-      putStrLn "🙅 These types have conflicts: "
-      traverse_ (\x -> putStrLn ("  " ++ Name.toString x)) types
+    showConflicts "terms" terms
+    showConflicts "types" types
+    where
+    terms    = R.dom $ Branch.termNamespace branch
+    types    = R.dom $ Branch.typeNamespace branch
+    showConflicts thingsName things =
+      when (not $ null things) $ do
+        putStrLn $ "🙅 These " <> thingsName <> " have conflicts: "
+        traverse_ (\x -> putStrLn ("  " ++ Name.toString x)) things
     -- TODO: Present conflicting TermEdits and TypeEdits
     -- if we ever allow users to edit hashes directly.
   FileChangeEvent _sourceName _src -> pure ()
@@ -503,16 +504,8 @@ notifyUser dir o = case o of
     if null allNames then []
     else [
       P.callout "❓" . P.sep "\n\n" . join $ [
-        if Set.null conflictedTypeNames then []
-        else [
-          P.wrap ("These" <> P.bold "types have conflicting definitions:")
-          `P.hang` P.commas (P.blue . prettyName <$> toList conflictedTypeNames)
-        ],
-        if Set.null conflictedTermNames then []
-        else [
-          P.wrap ("These" <> P.bold "terms have conflicting definitions:")
-            `P.hang` P.commas (P.blue . prettyName <$> toList conflictedTermNames)
-        ],
+        showConflictedNames "types" conflictedTypeNames,
+        showConflictedNames "terms" conflictedTermNames,
         [tip $ "This occurs when merging branches that both indepenently introduce the same name. Use "
             <> P.group ("`view " <> P.sep " " (prettyName <$> take 3 allNames) <> "`")
             <> "to see the conflicting defintions, then use `rename`"
@@ -521,6 +514,12 @@ notifyUser dir o = case o of
     ]
     where
       allNames = toList (conflictedTermNames <> conflictedTypeNames)
+      showConflictedNames things conflictedNames =
+        if Set.null conflictedNames then []
+        else [
+          P.wrap ("These" <> P.bold (things <> "have conflicting definitions:"))
+          `P.hang` P.commas (P.blue . prettyName <$> toList conflictedNames)
+        ]
 
   renderEditConflicts ppe (Branch.editConflicts -> cs@(e:_)) = [P.callout "❓" . P.sep "\n\n" $ [
     P.wrap $ "These" <> P.bold "definitions were edited differently"
@@ -676,246 +675,6 @@ emojiNote lead s = P.group (fromString lead) <> "\n" <> P.wrap s
 nothingTodo :: (ListLike s Char, IsString s) => P.Pretty s -> P.Pretty s
 nothingTodo s = emojiNote "😶" s
 
-type IsOptional = Bool
-
-data InputPattern = InputPattern
-  { patternName :: String
-  , aliases     :: [String]
-  , args        :: [(IsOptional, ArgumentType)]
-  , help        :: P.Pretty CT.ColorText
-  , parse       :: [String] -> Either (P.Pretty CT.ColorText) Input
-  }
-
-data ArgumentType = ArgumentType
-  { typeName :: String
-  , suggestions :: forall m v a . Monad m
-                => String
-                -> Codebase m v a
-                -> Branch
-                -> m [Line.Completion]
-  }
-
-showPatternHelp :: InputPattern -> P.Pretty CT.ColorText
-showPatternHelp i = P.lines [
-  P.bold (fromString $ patternName i) <> fromString
-    (if not . null $ aliases i
-     then " (or " <> intercalate ", " (aliases i) <> ")"
-     else ""),
-  help i ]
-
-validInputs :: [InputPattern]
-validInputs = validPatterns
- where
-  commandNames = patternName <$> validPatterns
-  commandMap   = Map.fromList (commandNames `zip` validPatterns)
-  helpPattern  = InputPattern
-    "help"
-    ["?"]
-    [(True, commandName)]
-    "`help` shows general help and `help <cmd>` shows help for one command."
-    (\case
-      []    -> Left $ intercalateMap "\n\n" showPatternHelp validPatterns
-      [cmd] -> case Map.lookup cmd commandMap of
-        Nothing  -> Left . warn $ "I don't know of that command. Try `help`."
-        Just pat -> Left $ help pat
-      _ -> Left $ warn "Use `help <cmd>` or `help`."
-    )
-  commandName =
-    ArgumentType "command" $ \q _ _ -> pure $ autoComplete q commandNames
-  branchArg = ArgumentType "branch" $ \q codebase _ -> do
-    branches <- Codebase.branches codebase
-    let bs = Text.unpack <$> branches
-    pure $ autoComplete q bs
-  definitionQueryArg = ArgumentType "definition query" $ \q _ b -> do
-    let names = HQ.toString <$> toList (Branch.allNamesHashQualified (Branch.head b))
-    pure $ autoComplete q names
-  noCompletions = ArgumentType "a word" $ \_ _ _ -> pure []
-  quit          = InputPattern
-    "quit"
-    ["exit"]
-    []
-    "Exits the Unison command line interface."
-    (\case
-      [] -> pure QuitI
-      _  -> Left "Use `quit`, `exit`, or <Ctrl-D> to quit."
-    )
-  validPatterns
-    = [ helpPattern
-      , InputPattern
-        "add"
-        []
-        []
-        (  P.wrap
-        $  "`add` adds to the codebase all the definitions from "
-        <> "the most recently typechecked file."
-        )
-        (\ws -> if not $ null ws
-          then Left $ warn "`add` doesn't take any arguments."
-          else pure $ SlurpFileI False
-        )
-      , InputPattern
-        "branch"
-        []
-        [(True, branchArg)]
-        (P.column2
-          [ ("`branch`", P.wrap "lists all branches in the codebase.")
-          , ( "`branch foo`"
-            , P.wrap
-            $  "switches to the branch named 'foo', "
-            <> "creating it first if it doesn't exist."
-            )
-          ]
-        )
-        (\case
-          []  -> pure ListBranchesI
-          [b] -> pure . SwitchBranchI $ Text.pack b
-          _ ->
-            Left
-              .  warn
-              .  P.wrap
-              $  "Use `branch` to list all branches "
-              <> "or `branch foo` to switch to or create the branch 'foo'."
-        )
-      , InputPattern
-        "fork"
-        []
-        [(False, branchArg)]
-        (  P.wrap
-        $  "`fork foo` creates the branch 'foo' "
-        <> "as a fork of the current branch."
-        )
-        (\case
-          [b] -> pure . ForkBranchI $ Text.pack b
-          _ ->
-            Left
-              .  warn
-              .  P.wrap
-              $  "Use `fork foo` to create the branch 'foo'"
-              <> "from the current branch."
-        )
-      , InputPattern
-        "find"
-        ["ls","list"]
-        [(True, definitionQueryArg)]
-        (P.column2
-          [ ("`find`", P.wrap $ "lists all definitions in the current branch.")
-          , ( "`find foo`"
-            , P.wrap
-            $  "lists all definitions with a name similar"
-            <> "to 'foo' in the current branch."
-            )
-          , ( "`find foo bar`"
-            , P.wrap
-            $  "lists all definitions with a name similar"
-            <> "to 'foo' or 'bar' in the current branch."
-            )
-          ]
-        )
-        (pure . SearchByNameI)
-      , InputPattern
-        "merge"
-        []
-        [(False, branchArg)]
-        (P.wrap "`merge foo` merges the branch 'foo' into the current branch.")
-        (\case
-          [b] -> pure . MergeBranchI $ Text.pack b
-          _ ->
-            Left
-              .  warn
-              .  P.wrap
-              $  "Use `merge foo` to merge the branch 'foo'"
-              <> "into the current branch."
-        )
-      , InputPattern "view"
-                     []
-                     [(False, definitionQueryArg)]
-                     (P.wrap "`view foo` prints the definition of `foo`.")
-                     (pure . ShowDefinitionI E.ConsoleLocation)
-      , InputPattern "edit"
-                     []
-                     [(False, definitionQueryArg)]
-                     (P.wrap "`edit foo` prepends the definition of `foo` to the top of the most recently saved file.")
-                     (pure . ShowDefinitionI E.LatestFileLocation)
-      , InputPattern
-        "rename"
-        ["mv"]
-        [(False, definitionQueryArg), (False, noCompletions)]
-        (P.wrap "`rename foo bar` renames `foo` to `bar`.")
-        (\case
-          [oldName, newName] -> Right $ RenameUnconflictedI
-            allTargets
-            (fromString oldName)
-            (fromString newName)
-          _ -> Left . P.warnCallout $ P.wrap
-            "`rename` takes two arguments, like `rename oldname newname`."
-        )
-      , InputPattern
-        "rename"
-        ["mv"]
-        [(False, definitionQueryArg), (False, noCompletions)]
-        (P.wrap "`rename foo bar` renames `foo` to `bar`.")
-        (\case
-          [oldName, newName] -> Right $ RenameUnconflictedI
-            allTargets
-            (fromString oldName)
-            (fromString newName)
-          _ -> Left . P.warnCallout $ P.wrap
-            "`rename` takes two arguments, like `rename oldname newname`."
-        )
-      , InputPattern
-        "alias"
-        ["cp"]
-        [(False, definitionQueryArg), (False, noCompletions)]
-        (P.wrap
-          "`alias foo bar` introduces `bar` with the same definition as `foo`."
-        )
-        (\case
-          [oldName, newName] -> Right $ AliasUnconflictedI
-            allTargets
-            (fromString oldName)
-            (fromString newName)
-          _ -> Left . warn $ P.wrap
-            "`alias` takes two arguments, like `alias oldname newname`."
-        )
-      , InputPattern
-        "update"
-        []
-        []
-        (  P.wrap
-        $  "`update` works like `add`, except "
-        <> "if a definition in the file "
-        <> "has the same name as an existing definition, the name gets updated "
-        <> "to point to the new definition. "
-        <> "If the old definition has any dependents, `update` will add "
-        <> "those dependents to a refactoring session."
-        )
-        (\ws -> if not $ null ws
-          then Left $ warn "`update` doesn't take any arguments."
-          else pure $ SlurpFileI True
-        )
-      , InputPattern
-        "propagate"
-        []
-        []
-        (P.wrap $ "`propagate` rewrites any definitions that"
-               <> "depend on definitions with type-preserving edits to use"
-               <> "the updated versions of these dependencies.")
-        (const $ pure PropagateI)
-      , InputPattern
-        "todo"
-        []
-        []
-        (P.wrap
-        $ "`todo` lists the work remaining in the current branch " <>
-          "to complete an ongoing refactoring."
-        )
-        (\ws -> if not $ null ws
-                   then Left $ warn "`todo` doesn't take any arguments."
-                   else pure $ TodoI)
-      , quit
-      ]
-  allTargets = Set.fromList [Names.TermName, Names.TypeName]
-
 completion :: String -> Line.Completion
 completion s = Line.Completion s s True
 
@@ -968,84 +727,3 @@ putPrettyLn' p = do
 getAvailableWidth :: IO Int
 getAvailableWidth =
   fromMaybe 80 . fmap (\s -> 100 `min` Terminal.width s) <$> Terminal.size
-
-getUserInput
-  :: (MonadIO m, Line.MonadException m)
-  => Map String InputPattern
-  -> Codebase m v a
-  -> Branch
-  -> BranchName
-  -> m Input
-getUserInput patterns codebase branch branchName = Line.runInputT settings $ do
-  line <- Line.getInputLine $
-    P.toANSI 80 (P.green (P.text branchName <> fromString prompt))
-  case line of
-    Nothing -> pure QuitI
-    Just l  -> case parseInput patterns $ words l of
-      Left msg -> lift $ do
-        liftIO $ putPrettyLn msg
-        getUserInput patterns codebase branch branchName
-      Right i -> pure i
- where
-  settings    = Line.Settings tabComplete (Just ".unisonHistory") True
-  tabComplete = Line.completeWordWithPrev Nothing " " $ \prev word ->
-    -- User hasn't finished a command name, complete from command names
-    if null prev
-      then pure $ autoComplete word (Map.keys patterns)
-    -- User has finished a command name; use completions for that command
-      else case words $ reverse prev of
-        h : t -> fromMaybe (pure []) $ do
-          p            <- Map.lookup h patterns
-          (_, argType) <- listToMaybe $ drop (length t) (args p)
-          pure $ suggestions argType word codebase branch
-        _ -> pure []
-
-main
-  :: forall v
-   . Var v
-  => FilePath
-  -> BranchName
-  -> Maybe FilePath
-  -> IO (Runtime v)
-  -> Codebase IO v Ann
-  -> IO ()
-main dir currentBranchName _initialFile startRuntime codebase = do
-  currentBranch <- Codebase.getBranch codebase currentBranchName
-  eventQueue    <- Q.newIO
-  currentBranch <- case currentBranch of
-    Nothing ->
-      Codebase.mergeBranch codebase currentBranchName Codebase.builtinBranch
-        <* (  putStrLn
-           $  "☝️  I found no branch named '"
-           <> Text.unpack currentBranchName
-           <> "' so I've created it for you."
-           )
-    Just b -> pure b
-  do
-    runtime                  <- startRuntime
-    branchRef                <- newIORef (currentBranch, currentBranchName)
-    cancelFileSystemWatch    <- watchFileSystem eventQueue dir
-    cancelWatchBranchUpdates <- watchBranchUpdates (readIORef branchRef) eventQueue codebase
-    let patternMap =
-          Map.fromList
-            $   validInputs
-            >>= (\p -> [(patternName p, p)] ++ ((, p) <$> aliases p))
-        getInput = do
-          (branch, branchName) <- readIORef branchRef
-          getUserInput patternMap codebase branch branchName
-    let awaitInput = do
-          -- Race the user input and file watch.
-          Async.race (atomically $ Q.peek eventQueue) getInput >>= \case
-            Left _ -> Left <$> atomically (Q.dequeue eventQueue)
-            x      -> pure x
-        cleanup = do
-          Runtime.terminate runtime
-          cancelFileSystemWatch
-          cancelWatchBranchUpdates
-    (`finally` cleanup)
-      $ E.commandLine awaitInput
-                           runtime
-                           (\b bn -> writeIORef branchRef (b, bn))
-                           (notifyUser dir)
-                           codebase
-      $ Actions.startLoop currentBranch currentBranchName
